requirements.txt
pip==8.1.1
Django<1.10
Markdown==2.6.6
celery==3.1.23
celery-haystack==0.10
django-celery-email==1.1.4
django-taggit==0.18.1
pytz==2016.4
requests==2.9.1
django-floppyforms==1.6.2
-e git://github.com/stephenmcd/django-overextends.git@master#egg=django-overextends
python-magic==0.4.6
django-tastypie==0.13.3
python-mimeparse>=0.1.4
django-configurations==1.0
django-storages==1.4.1
dj-database-url==0.4.1
django-cache-url==1.0.0
django-compressor==2.0
django-contrib-comments==1.7.0
unicodecsv==0.14.1
honcho==0.6.6
django-tinymce==2.3.0
anyjson==0.3.3
pysolr==3.5.0
python-docx==0.8.5

-e git://github.com/codeforhawaii/froide.git@master#egg=froide
-e git://github.com/django-haystack/django-haystack.git@f1ed18313777005dd77ed724ecbfb27c0b03cad8#egg=django-haystack

# -e git://github.com/stefanw/django-compressor.git@patch-1#egg=compressor
#psycopg2==2.6.1
#lxml==3.4.4
